TGP-1340: PR comments
