+++ app/connectors/GoodsRecordConnector.scala
@@ -19,7 +19,7 @@ package connectors
 import config.Service
 import models.{CategoryRecord, GoodsRecord}
 import models.router.requests.{CreateRecordRequest, UpdateRecordRequest}
-import models.router.responses.{CreateGoodsRecordResponse, GetGoodsRecordResponse}
+import models.router.responses.{CreateGoodsRecordResponse, GetGoodsRecordResponse, GetRecordsResponse}
 import org.apache.pekko.Done
 import play.api.Configuration
 import play.api.libs.json.Json
@@ -33,12 +33,21 @@ class GoodsRecordConnector @Inject() (config: Configuration, httpClient: HttpCli
   ec: ExecutionContext
 ) {
   private val tgpRouterBaseUrl: Service          = config.get[Service]("microservice.services.trader-goods-profiles-router")
+  private val dataStoreBaseUrl: Service          = config.get[Service]("microservice.services.trader-goods-profiles-data-store")
   private val clientIdHeader                     = ("X-Client-ID", "tgp-frontend")
   private def createGoodsRecordUrl(eori: String) =
     url"$tgpRouterBaseUrl/trader-goods-profiles-router/traders/$eori/records"
 
   private def singleGoodsRecordUrl(eori: String, recordId: String) =
     url"$tgpRouterBaseUrl/trader-goods-profiles-router/traders/$eori/records/$recordId"
+
+  private def getGoodsRecordsUrl(
+    eori: String,
+    lastUpdatedDate: Option[String] = None,
+    page: Option[Int] = None,
+    size: Option[Int] = None
+  ) =
+    url"$dataStoreBaseUrl/trader-goods-profiles-data-store/traders/$eori/records?lastUpdatedDate=$lastUpdatedDate&page=$page&size=$size"
 
   def submitGoodsRecord(goodsRecord: GoodsRecord)(implicit
     hc: HeaderCarrier
@@ -68,4 +77,16 @@ class GoodsRecordConnector @Inject() (config: Configuration, httpClient: HttpCli
       .setHeader(clientIdHeader)
       .execute[HttpResponse]
       .map(response => response.json.as[GetGoodsRecordResponse])
+
+  def getRecords(
+    eori: String,
+    lastUpdatedDate: Option[String] = None,
+    page: Option[Int] = None,
+    size: Option[Int] = None
+  )(implicit hc: HeaderCarrier): Future[GetRecordsResponse] =
+    httpClient
+      .get(getGoodsRecordsUrl(eori, lastUpdatedDate, page, size))
+      .setHeader(clientIdHeader)
+      .execute[HttpResponse]
+      .map(response => response.json.as[GetRecordsResponse])
 }
